backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
openai>=1.0.0
anthropic>=0.30.0
google-genai>=1.0.0
sse-starlette>=2.0.0
httpx>=0.25.0
